Use tmpdir instead of current dir for tempfile
The default value for the `:template` argument of
`cl-fad:open-temporary` (called by `cl-fad:with-output-to-temporary-file`) uses
a logical pathname host `temporary-files` which is initialized in a
system-dependent way.  This commit changes our call of
`cl-fad:with-output-to-temporary-file` to use that logical pathname host as
well.  The effect is that the tempfile will go into an appropriate tmpdir-like
location (e.g. /tmp) instead of being written to the current directory.

This also fixes #13.